requirements.txt
streamlit==1.39.0
autogluon.tabular==1.4.0
pandas>=1.5.0
numpy>=1.23.0
torch>=2.0.0
transformers>=4.44.0
sentencepiece>=0.1.99
sacremoses>=0.1.1
langdetect>=1.0.9
deepl>=1.19.0
catboost>=1.2.0,<2.0
lightgbm>=4.0.0
xgboost>=1.7.0
fastai>=2.7.0
